DBC ビジネスパッケージ　ビルドエラー修正

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShokanJuryoininKeiyakushaBuilder.java
@@ -84,7 +84,7 @@ public class ShokanJuryoininKeiyakushaBuilder {
      */
     public ShokanJuryoininKeiyakushaBuilder set履歴番号(Decimal 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
-        entity.setRirekiNo(履歴番号.intValue());
+        entity.setRirekiNo(履歴番号);
         return this;
     }
