+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShokanJuryoininKeiyakushaIdentifier.java
@@ -10,6 +10,7 @@ import java.io.Serializable;
 import jp.co.ndensan.reams.db.dbx.definition.valueobject.domain.HihokenshaNo;
 import jp.co.ndensan.reams.db.dbx.definition.valueobject.domain.ShoKisaiHokenshaNo;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 
 /**
  * 償還受領委任契約者の識別子です。
@@ -20,7 +21,7 @@ public class ShokanJuryoininKeiyakushaIdentifier implements Serializable {
     private final HihokenshaNo 被保険者番号;
     private final ShoKisaiHokenshaNo 証記載保険者番号;
     private final FlexibleDate 受付年月日;
-    private final int 履歴番号;
+    private final Decimal 履歴番号;
 
     /**
      * コンストラクタです。
@@ -33,7 +34,7 @@ public class ShokanJuryoininKeiyakushaIdentifier implements Serializable {
     public ShokanJuryoininKeiyakushaIdentifier(HihokenshaNo 被保険者番号,
             ShoKisaiHokenshaNo 証記載保険者番号,
             FlexibleDate 受付年月日,
-            int 履歴番号) {
+            Decimal 履歴番号) {
         this.被保険者番号 = 被保険者番号;
         this.証記載保険者番号 = 証記載保険者番号;
         this.受付年月日 = 受付年月日;
